Create first version
This is the first commit, where the base of the application is app.py
and in the templates folder, there is index.html. A .gitignore file has
also been added. The table "visitors" was created in PSQL.

This was used to test the functionality of the program locally on my
own computer, and it worked well. The home page includes a greeting
and a visitor counter.There is no other functionality yet.